移除多余字段
