+++ src/Pixeval/Controls/IllustratorView/IllustratorView.xaml
@@ -3,7 +3,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:card="using:Pixeval.Controls.Card"
     xmlns:local="using:Pixeval.Controls.IllustratorView"
-    xmlns:ui="using:Pixeval.Util.UI">
+    xmlns:markup="using:Pixeval.Util.MarkupExtensions">
     <Style TargetType="local:IllustratorView">
         <Setter Property="Template">
             <Setter.Value>
@@ -66,7 +66,7 @@
                                 Grid.Column="2"
                                 Width="20"
                                 VerticalAlignment="Stretch"
-                                Icon="{ui:FontIcon Glyph=ForwardE72A}" />
+                                Icon="{markup:FontIcon Glyph=ForwardE72A}" />
                         </Grid>
                     </card:CardControl>
                 </ControlTemplate>
