+++ src/Pixeval/Pages/Capability/SpotlightsPage.xaml
@@ -1,9 +1,8 @@
-﻿<controls:EnhancedPage
+<controls:EnhancedPage
     x:Class="Pixeval.Pages.Capability.SpotlightsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="using:Pixeval.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:illustratorContentViewer="using:Pixeval.UserControls.IllustratorContentViewer"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d" />
