+++ src/Pixeval/Pages/Download/DownloadListPage.xaml
@@ -3,14 +3,13 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:attributes="using:Pixeval.Attributes"
-    xmlns:card="using:Pixeval.Controls.Card"
     xmlns:controls="using:Pixeval.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:download="using:Pixeval.Pages.Download"
     xmlns:downloader="using:Pixeval.Download"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:toolkit="using:CommunityToolkit.WinUI.UI.Controls"
-    xmlns:ui="using:Pixeval.Util.UI"
+    xmlns:markup="using:Pixeval.Util.MarkupExtensions"
     mc:Ignorable="d">
     <Grid Padding="20,0">
         <Grid.RowDefinitions>
@@ -41,30 +40,30 @@
                 Background="Transparent"
                 DefaultLabelPosition="Right">
                 <AppBarButton
-                    Icon="{ui:FontIcon Glyph=CancelE711}"
+                    Icon="{markup:FontIcon Glyph=CancelE711}"
                     IsEnabled="{x:Bind _viewModel.IsAnyEntrySelected, Mode=OneWay}"
                     Label="{x:Bind _viewModel.SelectionLabel, Mode=OneWay}"
                     Tapped="CancelSelectButton_OnTapped" />
                 <AppBarButton
                     x:Uid="/DownloadListPage/SelectAllButton"
-                    Icon="{ui:FontIcon Glyph=SelectAllE8B3}"
+                    Icon="{markup:FontIcon Glyph=SelectAllE8B3}"
                     Tapped="SelectAllButton_OnTapped" />
                 <AppBarSeparator />
                 <AppBarButton
                     x:Uid="/DownloadListPage/ResumeButton"
-                    Icon="{ui:FontIcon Glyph=PlayE768}"
+                    Icon="{markup:FontIcon Glyph=PlayE768}"
                     Tapped="ResumeAllButton_OnTapped" />
                 <AppBarButton
                     x:Uid="/DownloadListPage/PauseButton"
-                    Icon="{ui:FontIcon Glyph=PauseE769}"
+                    Icon="{markup:FontIcon Glyph=PauseE769}"
                     Tapped="PauseAllButton_OnTapped" />
                 <AppBarButton
                     x:Uid="/DownloadListPage/StopButton"
-                    Icon="{ui:FontIcon Glyph=CancelE711}"
+                    Icon="{markup:FontIcon Glyph=CancelE711}"
                     Tapped="CancelAllButton_OnTapped" />
                 <AppBarButton
                     x:Uid="/DownloadListPage/DeleteButton"
-                    Icon="{ui:FontIcon Glyph=DeleteE74D}"
+                    Icon="{markup:FontIcon Glyph=DeleteE74D}"
                     Tapped="ClearDownloadListButton_OnTapped" />
             </CommandBar>
         </toolkit:DockPanel>
