+++ src/Pixeval/Pages/IllustratorViewer/IllustratorPage.xaml
@@ -5,7 +5,6 @@
     xmlns:controls="using:Pixeval.Controls"
     xmlns:controls1="using:CommunityToolkit.WinUI.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:Pixeval.Pages.Capability"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:userControls="using:Pixeval.UserControls"
     Loaded="IllustratorPage_OnLoaded"
