+++ src/Pixeval/Pages/MainPage.xaml
@@ -7,8 +7,8 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:pages="using:Pixeval.Pages"
-    xmlns:ui="using:Pixeval.Util.UI"
     xmlns:utilities="using:WinUI3Utilities"
+    xmlns:markup="using:Pixeval.Util.MarkupExtensions"
     mc:Ignorable="d">
     <controls:EnhancedWindowPage.Resources>
         <DataTemplate x:Key="IllustrationHeaderSuggestionTemplate">
@@ -129,7 +129,7 @@
                         Padding="0"
                         HorizontalAlignment="Right"
                         VerticalAlignment="Center"
-                        Content="{ui:FontIcon Glyph=SettingsE713,
+                        Content="{markup:FontIcon Glyph=SettingsE713,
                                               FontSize=13}"
                         Tapped="OpenSearchSettingButton_OnTapped" />
                     <Button
@@ -140,7 +140,7 @@
                         Padding="0"
                         HorizontalAlignment="Right"
                         VerticalAlignment="Center"
-                        Content="{ui:FontIcon Glyph=CameraE722,
+                        Content="{markup:FontIcon Glyph=CameraE722,
                                               FontSize=13}"
                         Tapped="ReverseSearchButton_OnTapped" />
                 </Grid>
@@ -148,44 +148,44 @@
             <NavigationView.MenuItems>
                 <NavigationViewItem
                     x:Uid="/MainPage/RecommendationsTab"
-                    Icon="{ui:FontIcon Glyph=CalendarE787}"
+                    Icon="{markup:FontIcon Glyph=CalendarE787}"
                     Tag="{x:Bind _viewModel.RecommendsTag}" />
                 <NavigationViewItem
                     x:Uid="/MainPage/RankingsTab"
-                    Icon="{ui:FontIcon Glyph=MarketEAFC}"
+                    Icon="{markup:FontIcon Glyph=MarketEAFC}"
                     Tag="{x:Bind _viewModel.RankingsTag}" />
                 <NavigationViewItem
                     x:Uid="/MainPage/BookmarksTab"
-                    Icon="{ui:FontIcon Glyph=HeartEB51}"
+                    Icon="{markup:FontIcon Glyph=HeartEB51}"
                     Tag="{x:Bind _viewModel.BookmarksTag}" />
                 <NavigationViewItem
                     x:Uid="/MainPage/FollowingsTab"
-                    Icon="{ui:FontIcon Glyph=PeopleE716}"
+                    Icon="{markup:FontIcon Glyph=PeopleE716}"
                     Tag="{x:Bind _viewModel.FollowingsTag}" />
                 <NavigationViewItem
                     x:Uid="/MainPage/SpotlightsTab"
-                    Icon="{ui:FontIcon Glyph=LibraryE8F1}"
+                    Icon="{markup:FontIcon Glyph=LibraryE8F1}"
                     Tag="{x:Bind _viewModel.SpotlightsTag}" />
-                <NavigationViewItem x:Uid="/MainPage/FeedsTab" Icon="{ui:FontIcon Glyph=LocationE81D}" />
+                <NavigationViewItem x:Uid="/MainPage/FeedsTab" Icon="{markup:FontIcon Glyph=LocationE81D}" />
                 <NavigationViewItem
                     x:Uid="/MainPage/RecentPostsTab"
-                    Icon="{ui:FontIcon Glyph=CaloriesECAD}"
+                    Icon="{markup:FontIcon Glyph=CaloriesECAD}"
                     Tag="{x:Bind _viewModel.RecentPostsTag}" />
                 <NavigationViewItemHeader x:Uid="/MainPage/DownloadAndHistoriesTab" />
                 <NavigationViewItem
                     x:Uid="/MainPage/HistoriesTab"
-                    Icon="{ui:FontIcon Glyph=HistoryE81C}"
+                    Icon="{markup:FontIcon Glyph=HistoryE81C}"
                     Tag="{x:Bind _viewModel.HistoriesTag}" />
                 <NavigationViewItem
                     x:Name="DownloadListTab"
                     x:Uid="/MainPage/DownloadListTab"
-                    Icon="{ui:FontIcon Glyph=DownloadE896}" />
+                    Icon="{markup:FontIcon Glyph=DownloadE896}" />
             </NavigationView.MenuItems>
             <NavigationView.FooterMenuItems>
-                <NavigationViewItem x:Uid="/MainPage/HelpTab" Icon="{ui:FontIcon Glyph=HelpE897}" />
+                <NavigationViewItem x:Uid="/MainPage/HelpTab" Icon="{markup:FontIcon Glyph=HelpE897}" />
                 <NavigationViewItem
                     x:Uid="/MainPage/AboutTab"
-                    Icon="{ui:FontIcon Glyph=RingerEA8F}"
+                    Icon="{markup:FontIcon Glyph=RingerEA8F}"
                     Tag="{x:Bind _viewModel.AboutTag}" />
                 <NavigationViewItem
                     x:Name="SettingsTab"
@@ -267,13 +267,13 @@
                     <AppBarButton
                         x:Uid="/MainPage/OpenSearchSettingButton"
                         Width="{StaticResource CollapsedAppBarButtonWidth}"
-                        Content="{ui:FontIcon Glyph=CameraE722,
+                        Content="{markup:FontIcon Glyph=CameraE722,
                                               FontSize=13}"
                         Tapped="ReverseSearchButton_OnTapped" />
                     <AppBarButton
                         x:Uid="/MainPage/ReverseSearchButton"
                         Width="{StaticResource CollapsedAppBarButtonWidth}"
-                        Content="{ui:FontIcon Glyph=SettingsE713,
+                        Content="{markup:FontIcon Glyph=SettingsE713,
                                               FontSize=13}"
                         Tapped="OpenSearchSettingButton_OnTapped" />
                 </CommandBar>
