+++ src/Pixeval/UserControls/IllustrationContainer.xaml
@@ -2,7 +2,6 @@
     x:Class="Pixeval.UserControls.IllustrationContainer"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="using:CommunityToolkit.WinUI.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:illustrationView="using:Pixeval.UserControls.IllustrationView"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
