+++ src/Pixeval/UserControls/IllustrationView/IllustrationViewModel.cs
@@ -53,9 +53,6 @@ public class IllustrationViewModel : IllustrateViewModel<Illustration>
 {
     private bool _isSelected;
 
-    public SoftwareBitmapSource? ThumbnailMediumSource => ThumbnailSources.TryGetValue(ThumbnailUrlOption.Medium, out var value) ? value : null;
-    public SoftwareBitmapSource? ThumbnailSquareMediumSource => ThumbnailSources.TryGetValue(ThumbnailUrlOption.SquareMedium, out var value) ? value : null;
-
     public Dictionary<ThumbnailUrlOption, SoftwareBitmapSource> ThumbnailSources { get; set; } = new();
 
     public IllustrationViewModel(Illustration illustration) : base(illustration)
@@ -151,13 +148,6 @@ public class IllustrationViewModel : IllustrateViewModel<Illustration>
         {
             ThumbnailSources[thumbnailUrlOption] = await stream.GetSoftwareBitmapSourceAsync(true);
             LoadingThumbnail = false;
-            OnPropertyChanged(thumbnailUrlOption switch
-            {
-                ThumbnailUrlOption.Medium => nameof(ThumbnailMediumSource),
-                ThumbnailUrlOption.SquareMedium => nameof(ThumbnailSquareMediumSource),
-                // ThumbnailUrlOption.Large => expr,
-                _ => WinUI3Utilities.ThrowHelper.ArgumentOutOfRange<ThumbnailUrlOption, string>(thumbnailUrlOption)
-            });
             OnPropertyChanged(nameof(ThumbnailSources));
             return true;
         }
@@ -170,13 +160,6 @@ public class IllustrationViewModel : IllustrateViewModel<Illustration>
             }
             ThumbnailSources[thumbnailUrlOption] = await ras.GetSoftwareBitmapSourceAsync(true);
             LoadingThumbnail = false;
-            OnPropertyChanged(thumbnailUrlOption switch
-            {
-                ThumbnailUrlOption.Medium => nameof(ThumbnailMediumSource),
-                ThumbnailUrlOption.SquareMedium => nameof(ThumbnailSquareMediumSource),
-                // ThumbnailUrlOption.Large => expr,
-                _ => WinUI3Utilities.ThrowHelper.ArgumentOutOfRange<ThumbnailUrlOption, string>(thumbnailUrlOption)
-            });
             OnPropertyChanged(nameof(ThumbnailSources));
             return true;
         }
