+++ src/Pixeval/UserControls/PixivReplyBar.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:iconButton="using:Pixeval.Controls.IconButton"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:ui="using:Pixeval.Util.UI"
+    xmlns:markup="using:Pixeval.Util.MarkupExtensions"
     Loaded="PixivReplyBar_OnLoaded"
     mc:Ignorable="d">
     <Grid>
@@ -19,7 +19,7 @@
             Grid.Column="0"
             HorizontalAlignment="Left"
             VerticalAlignment="Center"
-            Content="{ui:FontIcon Glyph=Emoji2E76E,
+            Content="{markup:FontIcon Glyph=Emoji2E76E,
                                   FontSize={StaticResource PixevalButtonIconSize}}"
             Style="{StaticResource PixevalRoundedButtonStyle}">
             <Button.Flyout>
@@ -33,12 +33,12 @@
                             <NavigationViewItem
                                 x:Name="EmojiButtonFlyoutEmojiSectionNavigationViewItem"
                                 x:Uid="/CommentRepliesBlock/EmojiButtonFlyoutEmojiSectionNavigationViewItem"
-                                Icon="{ui:FontIcon Glyph=Emoji2E76E}"
+                                Icon="{markup:FontIcon Glyph=Emoji2E76E}"
                                 IsSelected="True" />
                             <NavigationViewItem
                                 x:Name="EmojiButtonFlyoutStickersSectionNavigationViewItem"
                                 x:Uid="/CommentRepliesBlock/EmojiButtonFlyoutStickersSectionNavigationViewItem"
-                                Icon="{ui:FontIcon Glyph=Sticker2F4AA}" />
+                                Icon="{markup:FontIcon Glyph=Sticker2F4AA}" />
                         </NavigationView.MenuItems>
                         <Frame
                             x:Name="EmojiButtonFlyoutFrame"
@@ -65,7 +65,7 @@
             Height="30"
             HorizontalAlignment="Center"
             VerticalAlignment="Center"
-            Icon="{ui:FontIcon Glyph=SendE724,
+            Icon="{markup:FontIcon Glyph=SendE724,
                                FontSize={StaticResource PixevalSubscriptFontSize}}"
             Tapped="SendButton_OnTapped" />
     </Grid>
